Finishes error reworking.
